Melhorias realizadas no ProjetoController e PaticipacaoService

+++ aplicacoes/gpa/src/main/java/ufc/quixada/npi/gpa/service/PessoaService.java
@@ -40,5 +40,5 @@ public interface PessoaService {
 	
 	List<PessoaExterna> getAllPessoaExterna();
 	
-	PessoaExterna getPessoaExterna(int id);
+	PessoaExterna getPessoaExterna(Long id);
 }
