+++ aplicacoes/gpa/src/main/java/ufc/quixada/npi/gpa/service/impl/PessoaServiceImpl.java
@@ -162,7 +162,7 @@ public class PessoaServiceImpl implements PessoaService {
 	}
 
 	@Override
-	public PessoaExterna getPessoaExterna(int id) {
+	public PessoaExterna getPessoaExterna(Long id) {
 		return pessoaExternaRepository.find(PessoaExterna.class, id);
 	}
